requirements.txt
# FastAPI and server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.4

# Celery and Redis
celery>=5.3.6
redis>=5.0.1

# HTTP client
httpx>=0.27.0

# Web scraping
beautifulsoup4>=4.12.3
lxml>=5.1.0
crawlee[playwright]>=0.3.3
playwright>=1.45.0

# MongoDB drivers
pymongo>=4.6.1
dnspython>=2.4.2
motor>=3.3.2

# Email enrichment (100% free & open source)
python-whois>=0.9.4
email-validator>=2.1.0

# Utilities
aiofiles>=23.2.1
python-dotenv>=1.0.1

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Legacy Selenium scraper (optional — install manually if needed)
# selenium>=4.19.0
# selenium-wire>=5.1.0
# webdriver-manager>=4.0.1
